Fix combined discount derivation

+++ MightyLogic/HighGrowth/Strategies/__init__.py
@@ -11,10 +11,10 @@ from Heroes.OwnedHero import OwnedHero
 class HighGrowthStrategy:
     collection: Collection
     exclude: HeroSelector
-    gold_discount: Optional[int]
+    gold_discount: Optional[float]
 
-    def has_next(self, gold_remaining: Optional[int]) -> bool:
+    def has_next(self, gold_remaining: Optional[float]) -> bool:
         raise NotImplementedError("Method needs to be implemented by subclasses")
 
-    def process_next(self, gold_remaining: Optional[int]) -> Tuple[OwnedHero, LevelingSteps]:
+    def process_next(self, gold_remaining: Optional[float]) -> Tuple[OwnedHero, LevelingSteps]:
         raise NotImplementedError("Method needs to be implemented by subclasses")
